Add all tasks
Made a quite lot of changes. Changed the icons in the header and their
functionality. Now a user can add tasks when without specifying a
location to it.  Can view it under the second icon on map. Can view all
the task under one hood and manipulate them as required. Made the first
page as the main page where all the tasks are being listed rather the
map page.
Also made changes so that now, the tasks loads only once to the markers
when the program starts and keeps on updating. It just accesses all the
tasks from the database only once in the beginning.