Convert Array of Markers to Hash of Markers
By converting to hash, it will occupy less space, the google ids for the
marker used to start from values > 300 so by creating array we were also
creating 300 useless array spaces. But by using the hash, we avoid this
problem. Although now we can not know the numbers of markers being
stored, but it is anyways not being used anywhere.