Add info windows to marker and remove the set title and descripton from set-location
Now user need not go to set-location page to change the place name and
place description. One can click on the marker, an infowindow will pop
up showing the marker's place name and place description and the user
can edit it from there. Now on double clicking the tasks of the markers
are going to be shown.